Closes #51 (temptatively); changes tested flag
This probably needs some explanation: using -fPIC was no longer possible; so I went for some flag that could be used with ld and that could be testable. -s strips symbols. It's testable, since nm, a utility that should be installed wherever gcc is, looks for it. Output is affected by locale, which is why I simply use no, which should be good in Spanish (local to me) or English). That might fail somewhere else, but, well, we'll get there when we do. Probably just test for standard error, in that case